requirements.txt
gdown
numpy==1.23.1
scipy
ipdb
vtk
numpy-stl
joblib
opencv-python==4.7.0.72
tqdm
pyyaml
mujoco-py<2.2,>=2.1
scikit-image
wandb
gym
git+https://github.com/ZhengyiLuo/smplx.git@master
lxml
human_body_prior
autograd
chumpy
patchelf
chardet
pyopengl
torchgeometry
Scikit-learn
opencv-python
trimesh
scipy
matplotlib
roma
omegaconf